+++ example/android/app/src/main/kotlin/com/csx/flutter/flutter_hot_fix_csx_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example
+package com.csx.flutter.flutter_hot_fix_csx_example
 
 import io.flutter.embedding.android.FlutterActivity