+++ ios/Classes/FlutterHotFixCsxPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface FlutterHotFixCsxPlugin : NSObject<FlutterPlugin>
+@end