+++ ios/Classes/SwiftFlutterHotFixCsxPlugin.swift
@@ -1,0 +1,14 @@
+import Flutter
+import UIKit
+
+public class SwiftFlutterHotFixCsxPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "flutter_hot_fix_csx", binaryMessenger: registrar.messenger())
+    let instance = SwiftFlutterHotFixCsxPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+}