+++ lib/bsdiff/bsdiff/bsdiff.h
@@ -1,0 +1,18 @@
+//
+//  bsdiff.h
+//  DownloadTest
+//
+//  Created by 曹世鑫 on 2019/6/24.
+//  Copyright © 2019 曹世鑫. All rights reserved.
+//
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+    __attribute__((visibility("default"))) __attribute__((used)) int
+    Bsdiff_bsdiff(int argc, char *argv[]);
+#ifdef __cplusplus
+}
+#endif