+++ lib/flutter_hotfix_csx.dart
@@ -1,7 +1,0 @@
-library flutter_hotfix_csx;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}