+++ lib/helper/hotfix_helper.dart
@@ -1,0 +1,15 @@
+/*
+ * @Author: Cao Shixin
+ * @Date: 2021-06-28 10:37:35
+ * @LastEditors: Cao Shixin
+ * @LastEditTime: 2021-06-28 14:37:25
+ * @Description: 
+ */
+class HotFixHelper {
+  /*
+   * 当前可用资源的父级目录
+   */
+  static String currentValidResourceBasePath() {
+    return '';
+  }
+}