+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 0.0.1
+<!--
+ * @Author: Cao Shixin
+ * @Date: 2021-06-25 10:02:05
+ * @LastEditors: Cao Shixin
+ * @LastEditTime: 2021-06-25 10:18:46
+ * @Description: 
+-->
+## 0.0.1 【20210625】
 
-* TODO: Describe initial release.
+* 热更新本地资源支持.